Changed Add Entry form to also be the Edit Entry form.
Added new EntryMode enum to make the EntryForm for adding also be used for editing/viewing an entry.

Fixed the issue with the dummy form not properly hiding.  Used a 100 ms timer to "delay" calling Hide because apparrantly Show is being called somewhere after initialiazaion.

Added functionality for the following:
~~~Edit Entry
~~~View Entry
~~~Remove Entry
~~~Form Exit